+++ src/hash/partition.rs
@@ -21,10 +21,17 @@ pub struct GridPartitionPlugin<F = ()>(PhantomData<F>)
 where
     F: GridHashMapFilter;
 
-impl<F> Default for GridPartitionPlugin<F>
+impl<F> GridPartitionPlugin<F>
 where
     F: GridHashMapFilter,
 {
+    /// Create a new instance of [`GridPartitionPlugin`].
+    pub fn new() -> Self {
+        Self(PhantomData)
+    }
+}
+
+impl Default for GridPartitionPlugin<()> {
     fn default() -> Self {
         Self(PhantomData)
     }
@@ -284,7 +291,7 @@ where
             ComputeTaskPool::get(),
             None,
             |_, affected_cells| {
-                let _task_span = tracing::info_span!("parallel partition split").entered();
+                let _task_span = bevy_log::info_span!("parallel partition split").entered();
                 affected_cells
                     .iter_mut()
                     .filter_map(|(id, adjacent_hashes)| {